+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2018-10-26
+
+### Added
+
+- Support calculating the availability of a media object. #146
+
+### Changed
+
+- Update ObjectInterface to include more common functions #145
+
 ## [0.7.0] - 2018-10-09
 
 ### Added